+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Fixed: Bug where specific entries couldn't be openend
+* Added: `smtp.address` contains email-address of smtp data
+* Fixed: Email now supplies `From`-Header
 
 ## [v0.8.6] - 2018-11-20