+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.8.7] - 2018-11-21
+
 * Fixed: Bug where specific entries couldn't be openend
 * Added: `smtp.address` contains email-address of smtp data
 * Fixed: Email now supplies `From`-Header